fix: use export = for typescript module definition
